szalinski file comments
